SwiftMessagesの処理の切り出し
