+++ iOSEngineerCodeCheck/View/SearchViewController/SearchViewController.swift
@@ -101,38 +101,14 @@ extension SearchViewController: SearchPresenterOutput {
     }
 
     func showNothingEnteredAlert() {
-        let view = MessageView.viewFromNib(layout: .cardView)
-        view.configureTheme(.warning)
-        view.configureContent(title: "何も入力されていません", body: "")
-        view.button?.isHidden = true
-        view.layoutMarginAdditions = UIEdgeInsets(top: 20, left: 20, bottom: 20, right: 20)
-        (view.backgroundView as? CornerRoundingView)?.cornerRadius = 10
-        var config = SwiftMessages.Config()
-        config.presentationStyle = .center
-        SwiftMessages.show(config: config, view: view)
+        configureSearchResultsNotFountAlert()
     }
 
     func showNetworkErrorAlert() {
-        let view = MessageView.viewFromNib(layout: .cardView)
-        view.configureTheme(.error)
-        view.configureContent(title: "ネットワークエラーです", body: "")
-        view.button?.isHidden = true
-        view.layoutMarginAdditions = UIEdgeInsets(top: 20, left: 20, bottom: 20, right: 20)
-        (view.backgroundView as? CornerRoundingView)?.cornerRadius = 10
-        var config = SwiftMessages.Config()
-        config.presentationStyle = .center
-        SwiftMessages.show(config: config, view: view)
+        configureNetworkErrorAlert()
     }
 
     func showSearchResultsNotFountAlert() {
-        let view = MessageView.viewFromNib(layout: .cardView)
-        view.configureTheme(.warning)
-        view.configureContent(title: "検索結果がみつかりませんでした", body: "")
-        view.button?.isHidden = true
-        view.layoutMarginAdditions = UIEdgeInsets(top: 20, left: 20, bottom: 20, right: 20)
-        (view.backgroundView as? CornerRoundingView)?.cornerRadius = 10
-        var config = SwiftMessages.Config()
-        config.presentationStyle = .center
-        SwiftMessages.show(config: config, view: view)
+        configureSearchResultsNotFountAlert()
     }
 }
